test: re-organize test unit integration e2e
