+++ tests/unit/core/test_helpers.py
@@ -10,7 +10,11 @@ from tenacity import TryAgain, retry
 from tenacity.wait import wait_fixed
 
 from kili.core.graphql.operations.label.queries import LabelQuery
-from kili.core.helpers import RetryLongWaitWarner, format_result
+from kili.core.helpers import (
+    RetryLongWaitWarner,
+    format_result,
+    validate_category_search_query,
+)
 from kili.entrypoints.mutations.asset import MutationsAsset
 from kili.entrypoints.mutations.issue.helpers import get_labels_asset_ids_map
 from kili.exceptions import MissingArgumentError
@@ -219,3 +223,51 @@ class TestCheckWarnEmptyList(TestCase):
             ret = kili.change_asset_external_ids(new_external_ids=[], asset_ids=[])
         assert ret == []
         mocked__mutate_from_paginated_call.assert_not_called()
+
+
+@pytest.mark.parametrize(
+    "case, query, raise_error",
+    [
+        (
+            "user does not provide job",
+            "class_A.count > 0",
+            True,
+        ),
+        (
+            "user does not write count",
+            "JOB.class_A > 0",
+            True,
+        ),
+        (
+            "user can provide expressions separated by OR and AND",
+            "JOB.class_A.count > 0 AND JOB2.class_B.count == 3 OR JOB2.class_C.count < 3",
+            False,
+        ),
+        (
+            "user can provide expressions with parenthesis",
+            "JOB.class_A.count > 0 AND (JOB2.class_B.count == 3 OR JOB2.class_C.count < 3)",
+            False,
+        ),
+        (
+            "user can add spaces where it does not break the query",
+            "JOB. class_A.  count > 0 AND      JOB2.class_B. count ==   3",
+            False,
+        ),
+        (
+            "user did not close a parenthesis",
+            "JOB.class_A.count > 0 AND (JOB2.class_B.count == 3 OR JOB2.class_C.count < 3",
+            True,
+        ),
+        (
+            "user can have complex job and ategory name",
+            "JoB46_Hhzef*bzf66.class_Auzf657bdh----_.count > 0 ",
+            False,
+        ),
+    ],
+)
+def test_category_search_queries(case: str, query: str, raise_error: bool):
+    if raise_error:
+        with pytest.raises(ValueError):
+            validate_category_search_query(query)
+    else:
+        validate_category_search_query(query)
