+++ tests/unit/entrypoints/client/test_authentication.py
