+++ tests/unit/entrypoints/queries/asset/test_media_downloader.py
